backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
pytesseract==0.3.13
python-docx==1.1.2
Pillow==10.4.0
pdf2image==1.17.0
openpyxl==3.1.5
